feat: added usefetch hook for women data
