+++ src/pages/ProductPageWomen.jsx
@@ -4,54 +4,25 @@ import ProductFilters from "../components/ProductFilters";
 import ProductFooter from "../components/ProductFooter";
 import SecondaryNav from "../components/SecondaryNav";
 import DiscountHeader from "../components/DiscountHeader";
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import Spinner from "../components/Spinner";
 import { Suspense, lazy } from "react";
 import Loader from "../components/Loader";
 import ErrorMessage from "../components/ErrorComponent.jsx";
+import { useFetchWoMenData } from "../utils/fetchPaginatedData.js";
 
 const ProductDetails = lazy(() => import("../components/ProductDetails"));
 
-const ProductPageWomen = ({
-  womenOrgID,
-  womenApiKey,
-  womenAppID,
-  proxyUrl,
-}) => {
-  const [data, setData] = useState([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState(null);
+const ProductPageWomen = () => {
   const [page, setPage] = useState(1);
   const [progress, setProgress] = useState(0.3);
   const [currentNumber, setCurrentNumber] = useState(10);
+  const { data, error, loading } = useFetchWoMenData(page);
   const switchPage = (value, newNumber, newProgress) => {
     setPage(value);
     setCurrentNumber(newNumber);
     setProgress(newProgress);
   };
-
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        setLoading(true);
-        const response = await fetch(
-          `${proxyUrl}/products?organization_id=${womenOrgID}&reverse_sort=false&page=${page}&size=10&Appid=${womenAppID}&Apikey=${womenApiKey}`,
-        );
-        if (!response.ok) {
-          setError("Network error try again later");
-          throw new Error("Network response was not ok");
-        }
-        const result = await response.json();
-        setData(result);
-      } catch (error) {
-        setError("Network error try again later");
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchData();
-  }, [page]);
 
   if (loading) return <Spinner />;
   if (error) return <ErrorMessage message={error.message} />;
